requirements.txt
praw==7.8.1
prawcore==2.4.0
kafka-python==2.0.2
pyspark==3.5.3
